Added some memory cleaning, Vertex is a Subclass of Vector3d now.
This solution is temporary, I want to be able to use the same vertex multiple times without copying.
Implementing that in this current setup would be a headache (avoiding double frees) + I want to change
how the memory of meshes and triangles work, which means this kind of destructor implementation is
 unnecessary.
Vertex should have been a subclass from the start.